+++ internal/utils/strings.go
@@ -1,8 +1,0 @@
-package utils
-
-import "strings"
-
-// Contains checks if a string contains another string
-func Contains(s, substring string) bool {
-	return strings.Contains(s, substring)
-}